add emailVerified boolean back and move verification message to own page
