+++ client/lib/core/widgets/initial_loading_widget.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:client/core/routing/locations.dart';
 import 'package:client/core/data/services/logging_service.dart';
 import 'package:client/services.dart';
+import 'package:client/features/auth/presentation/views/verify_email_page.dart';
 import 'package:client/features/user/data/services/user_service.dart';
 import 'package:client/styles/styles.dart';
 import 'package:client/core/widgets/height_constained_text.dart';
@@ -96,11 +97,17 @@ class _InitialLoadingWidgetState extends State<InitialLoadingWidget> {
 
   @override
   Widget build(BuildContext context) {
-    if (context.watch<UserService>().signInState == SignInState.signedIn &&
-        _initialized) {
+    final userService = context.watch<UserService>();
+
+    if (userService.signInState == SignInState.signedIn && _initialized) {
+      final currentUser = userService.firebaseAuth.currentUser;
+      if (currentUser != null &&
+          !currentUser.isAnonymous &&
+          !currentUser.emailVerified) {
+        return VerifyEmailPage(email: currentUser.email ?? '');
+      }
       return widget.child;
-    } else if (context.watch<UserService>().signInState ==
-        SignInState.signedOut) {
+    } else if (userService.signInState == SignInState.signedOut) {
       return Scaffold(
         body: Center(
           child: HeightConstrainedText(
